chore: move employee page into its own component
